Allow users to block users that harass them
When you block a user:
* If they reply to a comment/post, you do NOT receive an orangered, and
* see NOTHING in your inbox.
* If they PM you, you do NOT receive an orangered, but the PM WILL show
* up in your inbox, with all fields replaced with the text "[blocked]".
* It's not readily possible to keep it out of the inbox while leaving it
* in the sender's sent box; and it needs to be in the sent box so that
* they can't tell that you've blocked them.
* You may not PM or make a comment reply to a user that you've blocked.
* This is to prevent abuse of the system by pre-emptively blocking a
* user and then sending them a series of harassment messages.

At present, the only way to block a user is from your inbox; if they PM
you or make a comment reply. There is a new "block user" button.
Unblocking a user can be done via /prefs/friends. This is to keep
'blocks' from being used commonly; in general, we prefer to encourage
the use of the downvote arrow for bad comments, and leave user-blocking
for true harassment scenarios.