Do API documentation formatting in template instead.
Changes the value of the api_methods nested dictionary to be a
dictionary instead of a string.  At a minimum, it will contain
an item named __doc__ with the docstring.  Any OAuth scopes,
instead of being prepended to the docstring, will be present as
an item named oauth2_scopes.  Additional introspected data (such
as required request arguments) can be added later.