Better date validation, do not redirect if form has errors
Now, in addition to checking that the date has the correct format (YYYY-MM)
we also check to make sure the date is in the future and the month is in
the range 01 to 12.

This fixes the problem where users were seeing a 500 server error after
entering invalid values for the month when entering their credit card info.

This change also fixes a problem where the form was redirecting even if there
were validation errors, leading the user to believe their authorization was
successful when it wasn't.