Record trans_id for freebies in campaign things
Initial campiagn Thing writes were recording an enum value for freebie
campaigns but it actually is useful to include the transaction id (which is
negative for freebies) because it references the freebie's record in the
bid table.

Note: This commit also includes a one-off script for fixing the trans_id
in existing campaign things.