Fix "what's this" help link in spotlight
When the user clicks "what's this?" we query the dom with .thing:visible to
find out what kind of link is being displayed so we know which blurb to show.
However, we were calling hide() on the link before doing the query, which
meant it didn't match the .thing:visible filter and we were always falling back
to the default.

This fixes a bug where we were showing the rising links help blurb when the
user clicked "what's this?" on a promoted link.