Always record trans_id in PromoCampaign things
Gets rid of default values so a data attribute will be saved even when trans_id
is zero. This change will make it possible to query for unpaid campaigns.

Also gets rid of payment state and makes sure the transaction id is preserved
even if the promoted link is rejected because this lets us link the campaign
to the refund in the promo log.