install-reddit: Allow domain to be configured at install time.
Automated runs of the installer need to be able to configure the
domain ahead of time to ensure that the site will come up in a valid
state. The install script will now check for an environment variable,
REDDIT_DOMAIN, and use that value if it is present. If not present,
it will default to the previous "reddit.local".