Add campaign thing name to tracking string
This change will make it possible to track traffic by campaign instead of just
by link.

Note that, as is, this change will cause the campaign fullname to be included
as part of the subreddit string in traffic_thingtarget and traffic_clicktarget,
two tables that are not currently used anywhere in the app.

To complete campaign traffic, the TargetedClickthroughsByCodename and
TargetedImpressionsByCodename in lib/traffic.py:make_report() will need to
be updated to store the campaign codename instead of link codename when
available.