Add support for "permanent" OAuth 2 grants.
If a client sends a user to the authorization page with a
"duration=permanent" parameter, we'll create a refresh token in addition
to the initial access token. When the client fetches the initial access
token with the authorization code, the refresh token will also be returned.
The client can then obtain new access tokens using the refresh token
instead of an authorization code.

Both refresh tokens and access tokens will be displayed on the user's apps
page (as well as the IP history page).