Available impressions viz on campaign edit page
Adds callback to the subreddit selector on the promo edit page that updates
an available inventory graph when the selected target changes. Shows available
front page inventory if no target is selected.

Right now this is set up in a new route. Eventually it should replace the old
promoted/edit_promo route.

TODO: change color of graph to show when inventory is running low
TODO: i18n of javascript strings