AllSR: Remove obviated sr_id > 0 rule.
This rule was added to ensure that links being returned had at least an
sr_id data attribute.  This prevented an issue where replication lag on
Postgres would cause half-baked objects to be returned by queries on the
slave and then loaded in that bizarro state into Memcached breaking them
for a long time.

Now that we are using proper transactions around Thing._commit, this is
no longer necessary and we can shed it for sanity and speed.