Gold Feature: "The Butler". Username monitoring in comments.
This is a new feature available only to Gold members. It defaults to on,
but may be turned off in preferences. When a username is mentioned on
the site in the /u/username format, and some conditions apply, the user
is notified by an orangered and a new type of item in the inbox. There
are several cases where notifications won't occur to reduce noise, see
butler.py for details.