Pre-fill reply box with quoted selected text
When "reply" is clicked, any text that the user has selected will be
pre-filled into the reply box, inside a blockquote. Allows a user to
select the text they're planning to reply to before clicking reply,
instead of needing to select, copy, paste, and add in "> " chars for
each line that they're quoting.