Subreddit settings: add ability to define a sticky
This allows subreddits to define a single self-post from inside the same
subreddit, which will be "stuck" to the top of the subreddit's hot page,
with distinct styling. Outside of viewing the subreddit's hot page
individually, the post will behave as normal. It also can only be hidden
by the user manually, and will not automatically hide due to being below
their score threshold, or being upvoted/downvoted with one of the user
preferences that normally hides posts they vote on.