Use a very conservative estimate of ad inventory.
Inventory is estimated using traffic to GET_listing. This is a pretty good
estimate based on historical trends (compare subreddit traffic to impressions for
promoted links targeted to the subreddit). It seems safer to start out with a low
estimate. This is especially needed because the CPM beta will only show promoted
links on the subreddit listing, not on multireddits or the front page.