Spam filter settings: add "all" level
This level makes it so that all items of the chosen type submitted to
the subreddit will be filtered initially. This is useful for subreddits
that want to be extremely strict with their submissions, as well as for
many subreddits to be able to go into "lockdown" in the case of a raid
or something similar.