Links: allow re-queuing for scrape if unsuccessful
If the scrape of a link fails for some reason, this will add a "retry
thumb" button to the link, which will re-add it to the queue to try
again. This will appear to similar people as the "nsfw" button - either
mods of the subreddit, the submitter, or the sponsor for a promoted
link.