/gold/creditgild: Check comment viewing permissions.
Previously, it was possible to generate a payment blob for a comment the
user does not have permission to view (in a private subreddit or
deleted) and then use the creditgild form to see the author and content
of the comment.  This adds a check to creditgild to ensure correct
permissions at display time.

This fixes an information disclosure vulnerability reported by Jordan
Milne (/u/largenocream).