Put media_domain warning on stderr to protect MR jobs.
For simple installs of reddit, g.domain == g.media_domain.  If this is
the case, the app will print a warning at startup.  However, this
warning was printed to stdout which meant that any jobs (such as the
mr_top stuff) using the output of a reddit process would get this
warning in their data streams. Bad times.