/buttonlite.js: Treat all requests as logged-out
buttonlite.js will attempt to link to an existing submission if one
already exists. This may leak whether a particular URL has been posted
to a private SR that the current user happens to have access to.

We normally circumvent this by setting c.user to UnloggedUser before
outputting anything in a format that cross-origin documents may read
(namely, JS.) buttonlite.js is a bit of a special case and marks itself
as outputting JS inside the actual controller, so we have to handle the
logging-out there too.