Stop displaying upvote/downvote numbers
This removes all displays of specific upvote/downvote numbers through
both the site and API, and only displays score.

Due to the API not previously including score for comments, all apps
were calculating the score themselves by doing ups - downs. So as to not
break everything, we now pass the score as the number of upvotes as
well, and always say that there are 0 downvotes. This should result in
still showing correct scores, but will make it look like nothing is ever
being downvoted in any apps that expose ups/downs.