Gold Checkout Redesign
This is a full redesign of gold checkout, including new resources for
gilding and creddits, at /gilding and /creddits. Work done in
collaboration by @powerlanguage and @umbrae.

This is primarily a visual refresh, although some flow has changed
(primarily that creddits now have their own endpoint, and there is a
/gold/payment page for finalizing payment which is handier for
analytics).