Show reports made by mods and users separately
This makes reports by moderators non-anonymous, and shows the name
of the moderator before their report reason. In addition, it shows the
number of reports with each reason for user reports, and orders them to
show the most common ones first, instead of the previous behavior of
arbitrary ordering (and possibly even not including the most common
reasons if there were a large number of different reasons).

This commit also adds a temporary deprecation warning for report_reasons
in the API, which should be removed in a month or so.