Twitter Cards: always show subreddit name
After looking at some usage with real posts, we determined that subreddit names
are pretty important.  So, we'll always show them now, and truncate the post
title accordingly.

Although subreddits can have fairly large names (20-something characters), most
of them don't, so in practice we shouldn't be cutting off much of the title.
Even in the worst case, it won't be so bad.

For a submission entitled "I don't even know how to describe the stupidity in
this gif...", before:

    I don't even know how to describe the stupidity in this gif... •...

and after:

    I don't even know how to describe the stupidity... • /r/reddit_test7