Messages: add full comments link where appropriate
Adding "full comments" links has long been a feature of RES.  In 2011, [we
added it natively to profile pages][0], but despite [the prediction][1], never
added it to inboxes as well.

Now we have!  If you view a comment reply or post reply message, you'll see a
nice shiny "full comment (n)" link, even without RES!

Per discussion with honestbleeps, the full-comments class exists solely to
allow RES to determine when not to add a duplicated link.

[0]: https://github.com/reddit/reddit/commit/eb46427a96351648ddbe7e7ad843e8a1a308532f
[1]: http://www.reddit.com/r/Enhancement/comments/kakch/announcement_seeing_double_full_comments_links/