Consistently use CSS prefixes
When browsers first introduce support for new CSS features, they often do so
with a *vendor prefix* - `-moz-` for Firefox, `-webkit-` for Chrome and Safari,
`-ms-` for IE.  It's annoying to deal with all of these, so we've written a few
Less functions to help.

Unfortunately, we haven't consistently used them.  This is most evident in how
the search box looks wonky to people using older versions of Firefox, because I
didn't use the `-moz-` prefix to `box-sizing`.

From a quick search, it looks like the only two functions we didn't use all the
time are `.transition` and `.box-sizing`; this should change all un-functioned
instances over.  This shouldn't affect a huge percentage of users, but it's
easy enough to do.

* http://caniuse.com/#search=transition
* http://caniuse.com/#search=box-sizing

Fixes #1146.