Search box: adjust tabindex
Adding a search submit button affected the tabbing behavior in the search box -
namely, it now began to take two tabs, rather than one, to get to the "limit my
search to /r/whatever" when on a subreddit page.

This changes add an explicit ordering: search text -> limit my search -> submit
button.  While this adds an extra tab to get to the submit button, that's not
really necessary, as pressing enter anywhere on the form also submits it.

There are many elements on the page with tabindices defined.  However, with the
exception of the "jump to content" link, they seem fairly random, so I'm not
terribly concerned about adjusted the overall index of the search form (that
is, how many tabs it takes to get to it from a freshly-loaded page).  It is
roughly in the same place as before - far enough down no one will want to tab
all the way to it. :)