Prefs: allow defaulting to new window opens
In general, non-power users tend to not open links in new tabs.  This provides
problems for sites like ours, since we end up with a very high bounce rate -
people click on a submission link, end up on some other interesting site, and
forget all about how they got there.  If that link instead opens in a new
window (tab), that problem is solved.

This is a pretty common behavior on link-aggregation sites, eg Google News.

We definitely don't want to change this for existing users, because that
creates a very confusing change.  Additionally, we're not sure if we will want
to keep this around for everyone, so rather than change the default and
backfill preferences, we're just setting it for anonymous users and new user
registrations - wrapped in a feature flag for easy alteration.