Redirect /r/ to /subreddits
This was [a user suggestion][0].  If you are familiar with directory paths,
you'd expect to see an index of subreddits here (one step up from each
subreddit page), but right now it's just a 404.

Technically it has not necessarily been a 404, since it will redirect to the
thread with id `r` if it exists.  However, this particular thread doesn't seem
to exist on the live site, and I don't think removing the short path to it is a
big deal, anyways.

[0]: http://www.reddit.com/r/ideasfortheadmins/comments/2jxqgb/why_is_wwwredditcomr_a_404_page_it_should_be_a/