Compact: extend scss compilation docs
This is the only scss file we have, and we don't change it often, so it's not
really integrated into the asset-building pipeline.  In the future, if we don't
get rid of it entirely, we'll probably convert it to less, but in the mean
time, let's make it slightly easier to work with.

I reverse-engineered these instructions, so they might not be exactly what
we've used previously.

In particular, I'm probably using a different version of the gems, which would
explain why the additional import was needed.