*VoteDetailsByDay: Page through votes to avoid timeouts when counting.
Since the number of votes is usually in the tens of millions currently,
get_count turned out to take far too long for reasonable RPC timeouts.
Instead of just boosting the timeout, this uses xget to page through the
columns instead and just counts them manually client side.
Unfortunately this is harder on C* and the network, but since we're only
using this in the awards stats job which runs daily it should be OK.