Flair selector: minor restyling
This is a tiny little touch-up of the flair selector box that just adds a
little more space between things, because I was annoyed at looking at it.  As
far as my choices for things:

* The `.flairselector h2` change was to make it consistent with the other
  dividing lines in the box.
* The button margin was chosen to be consistent with the "add" button on
  /about/moderators.

I decided not to mess with the wonky vertical alignment on the flair because
that gets changed by subreddits' customizations to flair display, so it's
probably not worth it to try and make it look better.