Prevent ban with custom message that will not send
A user will only be sent a ban notification if either of the following
is true:

1. they are subscribed to the subreddit
2. they have ever gained or lost karma (of any type) in the subreddit

Because of these conditions, there is some confusion in cases where
moderators ban someone that has made a post that didn't yet receive any
votes, where they still expect the user to receive a ban message and be
able to see the reason for the ban that they requested to include in it.

This commit makes it so that including a custom reason is not possible
in a case where the message won't actually be sent, to prevent that
confusion.