vote: Write LastModified in request rather than in queue.
This will allow us to trust LastModified for culling lookups to
prequeued_vote_keys. In a situation of bad queue lag, this could result
in more lookups going to Cassandra, as the LastModified culling in
DenomralizedRelation will be thrown off slightly, but that should be OK
as the bloom filters will catch the negative lookups quickly.