Comment threads: remove hot sort
Hot doesn't make much sense for comments because it was designed for threads.
There are very *very* few people who use it in comments, and we suspect most of
those are on accident - we don't provide good descriptions of what the sorts
are, so it's not obvious which of 'hot', 'best', and 'top' you should choose
when looking at a comment thread (particularly when 'hot' is the default for
submission listings).