Q&A sort: don't collapse comments on permalink pages
The comment-hiding behavior of Q&A sort has been confusing when viewing a
comment permalink, since it is fairly likely all the comments would be hidden.
This is particularly problematic when the user hasn't explicitly set the sort
for that page, eg by having it set as their default through the sticky sort
behavior or due to a subreddit- or thread-default sort.

Now we don't do any hiding of comments due to Q&A sort on permalink pages.