Contest mode: allow mods to sort however
Contest mode originally forced the sort to `top` for mods to assist in choosing
winners.  However, /u/timotab [pointed out][0] that the ability to use `new` as well
would be useful for moderation.

Now mods are able to use whichever sort method they wish when browsing a thread
in contest mode.

[0]: https://www.reddit.com/r/modnews/comments/2z913o/moderators_new_features_for_testing_before/cpgyzm7