Search: Add link to enable NSFW results
When the safe search feature is enabled, NSFW links and subreddits
will not appear in search results unless the over18 preference has
been set by the user.

For easy access to the over18 preference, add a link to enable NSFW
search results on the main search page and the subreddit search page.
Clicking on the link will redirect to the /over18 confirmation page.
The link will not appear if the over18 preference was already enabled.