Comment sort preference: dual-write to migrate
We ran the numbers on how many people have set some sort of sorting preference
(this includes *anything* using `VMenu`, since the field is serialized in the
database), and it's much more than I anticipated.  Per some discussion, we
decided we should make a good attempt to migrate over users' sort preferences
so as to not needlessly annoy them.

While we could write a migration script that iterates through all users with
this option set, that would take a long time and probably affect a lot of
long-disused accounts.  Instead we're opting to dual-write the new and old
preferences whenever they're updated through the preferences page (the new
way).  Additionally, we're checking for inconsistencies between those two
preferences on every view of a comments page, and setting the new preference to
the value of the old preference if any difference is found.  That allows us to
ship quickly while still invisibly migrating the preference for active users -
even those who are only viewing comment threads, not changing the sort.

This is likely going to lead to an increase in queries as we start to deploy
it, but it should be ok if we just take it slow.

After we've reached some point at which the number of accounts migrating over
is low enough we no longer care, we can rip out this commit.