Accept '/r/foo' everywhere 'foo' works
It's annoying to use the common subreddit syntax of `/r/foo` and be told you
can't do that.  We've already been normalizing this in a few places, but now we
should be accepting `/r/foo` or `r/foo` everywhere and silently stripping that
to just plain ol' `foo` for the rest of the code.

Here are the validators I found doing some sort of subreddit name check, and
where they're used:

- VSubredditName
    * multi.py
        + PUT_multi_subreddit - add a subreddit to a multi
- VAvailableSubredditName
    * api.py
        + POST_site_admin - create or configure a subreddit
- VMultiPath - already requires /r/
    * multi.py a bunch of places.
- VSubredditList - already doing this
- VSRByName
    * preferences.py - stylesheets everywhere
    * api.py
        + POST_compose - when sending a PM from a subreddit
    * multi.py
        + GET_list_sr_multis - getting a subreddit's multis
        + GET_multi_subreddit - get data about a subreddit in a multi
        + DELETE_multi_subreddit - remove a subreddit from a multi
- VSRByNames
    * api.py
        + GET_subreddit_recommendations
        + POST_rec_feedback - recommender feedback

One important thing to note: we don't want to just modify
`Subreddit.is_valid_name()` because that's used in lower-level code, like when
creating a `Subreddit` object, and that could cause all sorts of problems.