reddit themes: allow using of themes from 21-char subreddits
We mistakenly misread the name-enforcement regex `subreddit_rx` to say that
subreddit names can be a maximum of 20 characters, and so set the maximum
length of the field for custom subreddit themes at 20.  The maximum length is
actually 21, which means that a people have been prevented from using themes
from certain subreddits.