vote_q proc pickle -> JSON msg migration part 3
Part 3 of a multi-phase commit to migrate the message format
of the vote queues from pickle to JSON.

This removes the shim code in the vote queue processors, removing
references to the old pickle format. It should only be deployed
once the vote queues are fully clear of pickle formatted messages;
that is, after part 2 has been deployed AND the queue processors
have consumed any remaining pickle messages.