Make sure special pages can only be created by mods with config perms
Thanks to a report by @pokechu022. Previously we were only checking
if the user was a moderator with permission to moderate the wiki.

This didn't affect anything until the automoderator config page was
added, as all other special pages are autocreated and don't pass
the `may_not_create` check for _any_ user.