Search: Fix search query for empty multireddits
When searching within an empty multireddit, an empty OR clause
(i.e. `(or )`) was causing an uncaught InvalidQuery exception.
Do not generate an empty OR clause, and skip the search if the
subreddit query generation fails.

Removed extra ModContribSR condition, since ModContribSR is a
subclass of MultiReddit. Removed extra AllSR/DefaultSR condition,
since AllSR/DefaultSR is a subclass of FakeReddit.