AutoMod: obey user's pref for hiding flair
Previously, AutoModerator was using the internal values of a user's
flair, so it looked at what it was *set to*, regardless of whether it
was being publicly displayed or not. This changes back to the same
behavior as before it was integrated, where it will see no flair if the
user has chosen to disable their flair in the subreddit.