AutoMod: fix needs_media_data for compound checks
Previously, a rule would be considered to need media data if any checks
included a media check at all. For example, a check that looked at both
the submission's title and the media_title for a particular word would
not be applied to posts without media data. However, when in a compound
check like this, we should consider the media data optional, since there
are other non-media places that can be checked without it. This commit
does that, changing it so that a check will only cause the rule to
require media data if ALL of the fields being checked come from the
media, instead of if ANY of them do.