AutoMod: error if both "type: any" rules fail
If a rule definition specifies "type: any" (or does not specify type at
all), we try to create two separate rules for both comments and
submissions using the rule definition. If a particular type fails to
create, we just throw it away silently, considering this to mean that
the rule definition was actually only intended to apply to one type or
the other. However, if there is an error in the rule that results in
*both* the submission and comment attempts being thrown away, this means
that the entire rule will just end up being silently ignored. This adds
an error for that case, so that the user knows they need to fix their
definition.