Sharing: Handle narrow screens better.
When screens get narrow, the 'Share with:' label starts to wrap to two lines.
Since this is a translated string, we can't really just set a minimum width on
it and expect it to always work, but we can handle the wrapping case better.

This reduces the line-height to look less terrible, and keeps the labels
centered with the rest of the content when they need to wrap.