Traffic page: processing info in red only if slow
Previously the note about how traffic processing works would be red
whether it was currently slow or not. This makes it so that it's just
black and bold if it's the normal message about processing being on an
hourly basis, and red if it's the warning about things currently being
slow.