Add muting capability for modmail
Mods are able to mute a user from modmailing in for 24 hours.
They can click 'mute user' on the modmail of the sender they
want to mute or go to /about/muted to manage all of the muted
users.

If the user has interacted with the subreddit, they will be
messaged about the muting. If a subreddit or user try to
message each other when the user is muted, they will get an
error message and it won't go through.