Paper over negative comment counts
Sometimes, something somewhere goes wrong and we end up thinking there are a
negative number of comments in a thread.  This is obviously incorrect to users.

The full solution is to find out what's causing this and fix it, but in the
meantime, just changing those to 0 (what they probably should be) will fix the
display of this particular issue with the comment counts.