Timeouts: Send events on forbidden action attempts
VNotInTimeout checks to see if the user is in timeout. If the user is,
then an event is sent that the user attempted a forbidden action.

The details_text of the event is based off of the action_name. If the
action_name has been passed in, that will override the action_name that
exists in the request object.

Different actions have different targets, so the target chosen is based
off the action_name if the target hasn't been passed in. Additionally,
if the target is not of the type Subreddit, the subreddit that the action
is being performed in will be passed in the event payload.