EventV2: allow overriding of user data
Sometimes it's useful to be able to manually set things like the user id and
name, but previously `EventV2` would just overwrite those with the values from
the request.  Now we'll check each automatic value and only set it if it's not
already set in the data that was passed into the constructor.