setup.py: Remove check for distribute>=0.6.16.
This check was in place for 06c186c8c8056ee6be255fc6ee08bf121440a34a to make
sure that we were using a new enough version of distribute to handle
the "cythonize" command. distribute was a now-deprecated fork of setuptools,
and the versions of setuptools on precise and trusty are new enough to not
need this check.