Stop handling 304s for comments pages/userpages
We maintain a system that (poorly) tries to decide if it should send a
304 for logged-out users on comments pages and userpages. There are a
lot of errors in the logic for this system that's causing us to not send
304s in cases that we should, and send them in cases where we shouldn't.

Instead of trying to fix all the errors in this system, we're just going
to remove it and leave that up to the CDN. I performed a test on
2015-09-14 by completely disabling the system that confirmed that it
could be removed without significant negative impact to the site.

There will be behavior changes to /r/friends as a result of this.
Previously, we were taking advantage of the timestamps stored by this
system to try to figure out which 100 friends (out of the 500 most
recently added to friends list) had most recently interacted with the
site, and pull posts from those users for /r/friends. Since we no
longer have these timestamps, this has been switched to just pick 100
friends at random every half hour, similar to how subreddits are
selected for a user's front page.