More strongly separate defaults in top bar
If you are not subscribed to enough subreddits, the default/popular subs are
added to the top nav without any explanation, which strongly implies that you
are subscribed to those subs.  The separator between those and your
subscriptions was *almost* the same as the separator between subreddits, which
made it even less clear what was going on here.

Now we'll separate those sections with a pipe (|), like we do the other
sections.  The inclusion or not of the defaults is still a bit weird, but we
can deal with that another time.