Visually distinguish removed/deleted comments
If you're not an admin, mod, or the user whose content is being removed, a
comment that's been administratively removed has had the same appearance as a
comment the authoring user deleted, which is confusing.

Now, as a first step, comments removed by admins and mods will say "[removed]",
similar to self posts.