Don't rely on verdict for expunging self-post text
Currently, we will only expunge a self-post's text (replace it with
"[removed]") if the post has a "admin-removed" or "mod-removed" verdict.
This means that someone has to have *manually* removed it, so the text
of posts that were automatically removed by anti-spam measures is still
visible. This is not desirable when we have spam being correctly removed
via our automatic systems, so this changes it so that all removed posts
will have their self-text hidden, regardless of what removed them.