/r/friends: use friends that posted most recently
This takes advantage of the new last_submit_time and last_comment_time
timestamps on the Account objects to determine which friends have posted
most recently.

Ideally the ability to sort on data columns would have been added into
the general _query() function, but that seems to be fairly complex, so
this was a lot simpler of a method for addressing this specific problem.