SubredditRules: Additional API errors.
Adds the following errors:

SR_RULE_EXISTS if trying to create a new rule with a short_name that
already exists.
SR_RULE_DOESNT_EXIST if trying to update a rule with a short_name that
does not exist.
SR_RULE_TOO_MANY if trying to create a new rule when the subreddit
already has the max number of rules.