Make mods in timeout send PMs as themselves
When a moderator visits the PM composition page, we change the 'from' field a
bit depending on how a few things.  For a moderator who isn't in timeout,

* /message/compose shows a drop-down of their username and all subreddits they
  moderate
* /r/mod/message/compose shows a drop-down of their username and all subreddits
  they moderate
* /r/example/message/compose shows a read-only field indicating the message is
  coming from /r/example, unless the user doesn't moderate that subreddit (in
  which case they get a 403)

None of this behavior should be altered by this commit.

When a user is in timeout, we only allow them to message /r/reddit.com.  While
there are separate backend restrictions for this, we also have the frontend
force only that option.  However, moderators have been getting one of their
subreddits forced as the sender instead, which causes an error (subreddits
cannot message other subreddits) and doesn't really make much sense, anyways.

I refactored the logic here a bit to make this all a bit more straightforward.
In particular, the only time `from_user` was set to false was in the second
case above, and `mod_srs` was set at the same time to a one-element list, which
lead to some strange length-checking to make sure the one-subreddit case wasn't
because of only moderating one subreddit.

Now a moderator in timeout will no longer have the option to choose who to send
a message as - it will look like the non-mod message compose page, which
defaults to sending a message as the user themselves.