Inbox: fix 'mark all as read' error case
After the 'mark all as read' button fires off a POST to request the
marking-of-all-as-read, it polls `/api/me.json` to find out whether or not the
user has any mail, refreshing the page when they no longer do.  As a fail-safe,
if the poll fails 20 times, we just refresh the page anyways, assuming it's
probably a race condition and there's a new message waiting for them.

However, the first call to the polling function was coming directly as a
callback from `r.ajax`, which meant it was passed in the response instead of a
null count (as it expects).  Thanks to Javascript's weak typing, the counter
increment was instead just appending the string "1" to the end of the response,
which would never trigger the end condition.

The callback is now a simple little wrapper around the polling function to
prevent any arguments from being initially passed in.