CommentTreeStorageV1: Don't trigger a rebuild on miss
Instead just assume that the comment tree is empty. One downside
of this approach is that until a comment tree is created (when
a new comment is added), each attempt to retrieve the comment tree
will result in a read hitting Cassandra (permacache).

Doing a rebuild was bad because it caused lock contention if there
were simultaneous attempts to:

1) View the link's comments page
2) Add a new comment
3) Vote on a comment (update_comment_votes retrieves the tree for qa sort)