Create subreddit button: hide if user can't create
We only show the create subreddit button in the sidebar to users whose
account is at least 30 days old, but since there are now some additional
requirements for creating a subreddit, this means that sometimes we end
up showing the button to people that won't be able to successfully
create a subreddit. This keeps the 30-day restriction, but on top of
that also checks that the user's actually going to be able to create
before showing the button.