Subreddit Rules: Fix broken link to rules page
The string formatting inside of the 'subreddit_report_form' template function is causing
the underscore template tag to get escaped, so the link doesn't render properly on the
frontend. Moving the call to 'unsafe' fixes it, and should be safe since 'sr_name' will only ever be the name of a subreddit or the underscore template tag.