Stop checking title for NSFW/NSFL at display-time
As of 1148446f1463f0a2bd13397565a39a0f6266d263 we started checking the
title of submissions at submit-time to see if they should be set to
NSFW. Between that and backfilling all existing submissions to have the
NSFW flag set correctly based on their titles, we can now remove the
display-time check.

This will allow moderators and submitters to turn off NSFW even if
"NSFW" or "NSFL" is in the title. This was an annoyance in cases where
the submitter puts something like "(not NSFW)" in the title but then had
the flag stuck on because of the display-time check.