Only redirect OAuth authorize requests as a result of user action.
Redirecting to the client's `redirect_uri` for any error was an
open redirect. Only redirecting as a result of user action ensures
we're checking errors and only sending a resource owner to a new
page if they've seen the authorize page and have taken some action.

Thanks to /u/avlidienbrunn for reporting!