Self posts: start giving karma
This commit starts giving and tracking karma for self-posts as well as
link posts. The karma is stored separately internally, but displayed and
otherwise treated as part of link karma. This will allow us to easily
revert this change and go back to the original link-only karma values if
desired.