POST_block: fix inbox-check loophole for mods
The "is this item in the user's inbox?" check is supposed to be skipped
if the item is a modmail to a subreddit that the user moderates, but the
previous implementation of it also allowed mods to block from any
comment in a subreddit the user moderates.

This change updates the check to make sure that it's a message as well.