Fix context-based Link.stickied value in API
The `stickied` attribute was previously dependent on the current
subreddit context, meaning that it would only be true when the request
was "inside" the same subreddit as the Link is stickied in. This is good
when it's being used for styling purposes since we don't want to style
sticky posts in places like the front page, but it also meant that the
value was incorrect in the API when using endpoints like /api/info where
you're not in the subreddit's context.

This change adds a new `use_sticky_style` onto wrapped Links that has
the same behavior as `stickied` previously did, and uses that in
templates for styling decisions. The `stickied` attribute now reflects
whether the Link is stickied in its subreddit, regardless of the context
that the info is being displayed in.