OAuth2Client: fix definition for .deleted
The deleted property previously wasn't defined as a bool, so setting it
to True or False ended up with the strings "True" or "False" being saved
to the backing CF. This meant that trying to explicitly set a client as
not deleted would only work if you set the value as an empty string (or
completely removed the property), since attempting to set it as False
would save a string that still evaluates as True in a boolean context.

As part of this change, I also ran a backfill script that went through
and converted all existing .deleted values in the database. Anything
with a value of "True" was converted to "1" (since that's the only value
that something defined in _bool_props will accept as True), and it just
removed the property from anything with empty strings as values.

In addition, this adds .deleted to OAuth2Client._defaults so that it's
not necessary to use getattr() all the time to access it.