LinksByUrlAndSubreddit: truncate very long urls
There are a few urls in the system that are extremely long (100,000+
chars). Cassandra's maximum key length is 64kb, so the whole url can't
be used as a key for these. This truncates the url at 65,000 characters
to prevent that from happening.

We should probably also implement a maximum url length, but this
prevents crashes from the pre-existing urls longer than that.