Remove multiple-sticky transitional code
Subreddit sticky posts (announcements) were originally implemented with
only allowing a single sticky to be set. When we added support for a
second sticky, there was some transitional code added that would handle
automatically converting the old single sticky to the new storage format
for multiple.

Everything's been long converted now, so this is no longer necessary to
keep around.