search: Disable faceting
Faceting returns counts of query matches per subreddit name so we can
show the user all the subreddits that contain links for their query.
This ends up returning many unpopular subreddits, which when looked up
by name are often missing from cache and result in a postgres query. This
is bad for performance, and the feature is not popular with users so it is
safe to remove.